apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaoy-mdhub-api"
version = "1.0.0"
description = "Backend API for Local High-Availability Market Data Hub"
authors = [
    {name = "John, Product Manager", email = "john@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["market-data", "trading", "high-availability", "fastapi", "vnpy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "ta-lib==0.6.4",
    "vnpy==4.0.0",
    "vnpy_ctp==6.7.7.2",
    "vnpy_sopt==3.7.1.0",
    "numpy>=2.2.3",
    "pandas==2.3.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-multipart==0.0.6",
    "typing-extensions==4.14.0",
    "sqlalchemy==2.0.23",
    "pymysql==1.1.0",
    "pymongo==4.6.0",
    "alembic==1.13.1",
    "pyzmq==26.3.0",
    "msgpack==1.0.7",
    "websockets==12.0",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "PySide6==6.9.1",
    "PySide6-Essentials==6.9.1",
    "PySide6-Addons==6.9.1",
    "shiboken6==6.9.1",
    "pyqtgraph==0.13.7",
    "qtpy==2.4.3",
    "qdarkstyle==3.2.3",
    "plotly==6.1.2",
    "tqdm==4.67.1",
    "pytz==2025.2",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "python-dateutil==2.9.0.post0",
    "structlog==23.2.0",
    "prometheus-client==0.19.0",
    "loguru==0.7.3",
    "annotated-types==0.7.0",
    "anyio==3.7.1",
    "attrs==25.3.0",
    "bcrypt==4.3.0",
    "cffi==1.17.1",
    "click==8.2.1",
    "cryptography==45.0.4",
    "deap==1.4.3",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "fastjsonschema==2.21.1",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httptools==0.6.4",
    "idna==3.10",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-core==5.8.1",
    "Mako==1.3.10",
    "MarkupSafe==3.0.2",
    "narwhals==1.44.0",
    "nbformat==5.10.4",
    "packaging==25.0",
    "platformdirs==4.3.8",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.27.0",
    "traitlets==5.14.3",
    "typing-inspection==0.4.1",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.1",
]

[project.urls]
"Homepage" = "https://github.com/xiaoy/xiaoy_mdhub"
"Repository" = "https://github.com/xiaoy/xiaoy_mdhub.git"
"Documentation" = "https://github.com/xiaoy/xiaoy_mdhub/docs"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"